Use virtual sizes to centralize image
Use values provided vinfo.{x,y}res_virtual in order to put the image in
the center of secondary monitors. In this case, the virtual size is the
resolution o the biggest one. We can then use xoffset and yoffset to
tell which are should be visible.

It needs proper testing.