Hide cursor before going to graphics mode
If we don't hide the cursor, it can happen that it will remain drawn
because we don't clear the screen nor set a background color. Rather
than memsetting all to 0 or another background color, we do this little
hack to disable the cursor. VGA console has TERM=linux, so this works,
but it could print weird things if user redirected the console to
another tty.