genstaticlogo: Allow the creation of more than one static image
Multiple static images are useful for coming animations while computer
is booting. In case genstaticlogo receives only one image as argument it
will generate a static image struct containing the information of that
image (case 1 below). If more than one image is given, it will rather
generate several structs describing the images and one struct as the
array with pointers to each image (case 2 below).

For example:

1)

 $ src/genstaticlogo dietsplash_staticlogo logo.h logo.ppm

It generates a file named logo.h containing a

 static struct image dietsplash_staticlogo

initialized with data obtained from image.

2)

 $ src/genstaticlogo dietsplash_staticanim anim.h anim-01.ppm anim-02.ppm

It generates a file named anim.h containing:

 static struct image dietsplash_staticanim0
 static struct image dietsplash_staticanim1

filled as in case 1. Besides, it generates the following array:

 static struct image *dietsplash_staticanim[] {
	 &dietsplash_staticanim0,
	 &dietsplash_staticanim1
 }

So, in dietsplash we would use the elements of the array above to change
the images in an animation.